Revert "Refactor Duke to MeetingJio"
This reverts commit 6d13608556041d2e6ca6e5dcb0af2ef9e4f275a6.
