+++ src/test/java/seedu/duke/DukeTest.java
@@ -1,10 +1,10 @@
-package seedu.meetingjio;
+package seedu.duke;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class MeetingJioTest {
+class DukeTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
